+++ Runtime/Main/Localization.cs
@@ -2,19 +2,39 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using Newtonsoft.Json;
-using SimplyLocalize.Runtime.Data;
-using SimplyLocalize.Runtime.Data.Keys;
-using SimplyLocalize.Runtime.Data.Keys.Generated;
 using UnityEngine;
+using Object = UnityEngine.Object;
 
-namespace SimplyLocalize.Runtime.Main
+namespace SimplyLocalize
 {
     public static class Localization
     {
+        private static LocalizationKeysData _localizationKeysData;
         private static Dictionary<string, Dictionary<string, string>> _allLocalizations;
+        private static Dictionary<string, Dictionary<Object, Object>> _allLocalizationsObjects;
+        
         private static Dictionary<string, string> _currentLocalization;
+        private static Dictionary<Object, Object> _currentLocalizationObjects;
 
         private static FontHolder _fontHolder;
+
+        public static LocalizationKeysData LocalizationKeysData
+        {
+            get
+            {
+                if (_localizationKeysData != null) return _localizationKeysData;
+                
+                if (!TryLoadDefaultLanguage())
+                {
+                    throw new NotImplementedException(
+                        $"{nameof(AllLocalizations)} is empty. " +
+                        $"You should call {nameof(Localization)}.{nameof(SetLocalization)} method before use localization components " +
+                        $"or set default localization data in Localization Settings window.");
+                }
+                
+                return _localizationKeysData;
+            }
+        }
         
         public static Dictionary<string, Dictionary<string, string>> AllLocalizations
         {
@@ -27,10 +47,28 @@ namespace SimplyLocalize.Runtime.Main
                     throw new NotImplementedException(
                         $"{nameof(AllLocalizations)} is empty. " +
                         $"You should call {nameof(Localization)}.{nameof(SetLocalization)} method before use localization components " +
-                        $"or set default localization data in {nameof(LocalizationKeysData)} asset.");
+                        $"or set default localization data in Localization Settings window.");
                 }
 
                 return _allLocalizations;
+            }
+        }
+        
+        public static Dictionary<string, Dictionary<Object, Object>> AllLocalizationsObjects
+        {
+            get
+            {
+                if (_allLocalizationsObjects != null) return _allLocalizationsObjects;
+
+                if (!TryLoadDefaultLanguage())
+                {
+                    throw new NotImplementedException(
+                        $"{nameof(AllLocalizations)} is empty. " +
+                        $"You should call {nameof(Localization)}.{nameof(SetLocalization)} method before use localization components " +
+                        $"or set default localization data in Localization Settings window.");
+                }
+
+                return _allLocalizationsObjects;
             }
         }
 
@@ -46,7 +84,21 @@ namespace SimplyLocalize.Runtime.Main
             }
         }
 
+        public static Dictionary<Object, Object> CurrentLocalizationObjects
+        {
+            get
+            {
+                if (_currentLocalizationObjects != null) return _currentLocalizationObjects;
+                
+                AllLocalizationsObjects.TryGetValue(CurrentLanguage, out _currentLocalizationObjects);
+
+                return _currentLocalizationObjects;
+            }
+        }
+
         public static string CurrentLanguage { get; private set; }
+        public static bool EnableLogging => LocalizationKeysData.EnableLogging;
+        public static bool LoggingOnlyInEditor => LocalizationKeysData.LoggingInEditorOnly;
 
         public static event Action LanguageChanged;
 
@@ -61,14 +113,21 @@ namespace SimplyLocalize.Runtime.Main
             
             if (localization == null)
             {
-                Debug.LogWarning($"{nameof(localization)} not founded");
+                Logging.Log($"{nameof(localization)} not founded", LogType.Warning);
                 return false;
             }
 
             var allLocalizations = JsonConvert.DeserializeObject<Dictionary<string, Dictionary<string, string>>>(localization.text);
             allLocalizations.TryGetValue(localizationData.i18nLang, out _currentLocalization);
+            
+            var allLocalizationsObjects = _localizationKeysData.ObjectsTranslations
+                .ToDictionary(d => d.Key, d => d.Value
+                    .ToDictionary(d => d.Key, d => d.Value));
 
+            allLocalizationsObjects.TryGetValue(localizationData.i18nLang, out _currentLocalizationObjects);
+            
             _allLocalizations = allLocalizations;
+            _allLocalizationsObjects = allLocalizationsObjects;
             _fontHolder = localizationData.OverrideFontAsset;
             
             if (CurrentLanguage == null)
@@ -91,7 +150,7 @@ namespace SimplyLocalize.Runtime.Main
             
             if (localizationResource == null)
             {
-                Debug.LogWarning($"{nameof(localizationResource)} not founded");
+                Logging.Log($"{nameof(localizationResource)} not founded", LogType.Warning);
                 return false;
             }
 
@@ -118,13 +177,14 @@ namespace SimplyLocalize.Runtime.Main
 
         private static bool TryLoadDefaultLanguage()
         {
-            var data = Resources.Load<LocalizationKeysData>(nameof(LocalizationKeysData));
-            if (data.DefaultLocalizationData != null)
+            _localizationKeysData ??= Resources.Load<LocalizationKeysData>(nameof(LocalizationKeysData));
+            
+            if (_localizationKeysData.DefaultLocalizationData != null)
             {
-                SetLocalization(data.DefaultLocalizationData);
+                SetLocalization(_localizationKeysData.DefaultLocalizationData);
 
-                Debug.Log($"Used default language: {CurrentLanguage}, please call the method: " +
-                          $"{nameof(Localization)}.{nameof(SetLocalization)} to use another one.");
+                Logging.Log($"Used default language: {CurrentLanguage}, please call the method: " +
+                            $"{nameof(Localization)}.{nameof(SetLocalization)} to use another one.");
 
                 return true;
             }